v0.0.1.0, first set of emoji "components" added.
Emojis are built using components. There are three categories: face, eyes, and mouth. In most cases, one of each is required to make an emoji.
This commit adds the first batch of components. There is the basic yellow face, a couple pairs of eyes, and a couple mouths.
The actual emojis have not been added yet; that is coming next version.
I call this version "Phase 1", in which components are added. "Phase 2" will be done once all combinations of emojis that can be created using these components have had recipes created for them.