Delete run-data directory
This follows the last commit and deletes the entire run-data directory as the configuration contained within appears to only be needed for development. Hopefully this is the last thing I need to delete from GitHub in order to clear to repo of unneeded stuff. Next commit will contain changes to the .gitignore to complement this deletion of files.